% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{crossclassified_data}
\alias{crossclassified_data}
\title{Sample cross-classified data for analysis}
\format{
Sample data contains:
 \describe{
   \item{f1id}{factor-1 cluster (e.g., school) identifier}
   \item{f1id}{factor-2 cluster (e.g., neighborhood) identifier}
   \item{f12id}{combined factor-12 cluster (e.g., interaction of school and neighborhood) identifier}
   \item{Y}{continusous outcome}
   \item{Z}{binary treatment (1=treated, 0=control)}
   \item{X1}{first individual-level confounder}
   \item{X2}{second individual-level confounder}
   \item{X3}{third individual-level confounder}
   \item{W1}{first factor-1 confounder (e.g., school climate)}
   \item{W2}{second factor-1 confounder, considered as omitted/measured}
   \item{Q1}{first factor-2 confounder (e.g., perceived negiborhood safety)}
   \item{Q2}{second factor-2 confounder, considered as omitted/measured}
   \item{lps}{true propensity score logit}
   \item{ps}{true propensity score}
   \item{Y1}{true potential treatment outcome}
   \item{Y0}{true potential control outcome}
 }
}
\usage{
crossclassified_data
}
\description{
Sample cross-classified data where individuals (e.g., students) belong to two clusters simultaneously (e.g., schools and neighborhoods)
}
\keyword{datasets}
