% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combFun.R
\name{CUClusterML.setup}
\alias{CUClusterML.setup}
\title{Setup function}
\usage{
CUClusterML.setup(...)
}
\arguments{
\item{...}{additional arguments ...}
}
\value{

}
\description{
setup function to implement CURobustML
}
\examples{
\dontrun{
CUClusterML.setup()
}
}
