% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combFun.R
\name{CURobustML.setup}
\alias{CURobustML.setup}
\title{Setup function}
\usage{
CURobustML.setup(...)
}
\arguments{
\item{...}{additional arguments ...}
}
\value{

}
\description{
setup function to implement CURobustML
}
\examples{
\dontrun{
CURobustML.setup()
}
}
