% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combFun.R
\name{DR}
\alias{DR}
\title{Doubly robust estimator}
\usage{
DR(Y, Z, interZ = formula(~1), Z.hat, Y1.hat, Y0.hat, data)
}
\arguments{
\item{Y}{continuous outcome variable}

\item{Z}{binary treatment indicator, 1 - treatment, 0 - control}

\item{interZ}{formula that contains the variables that "interact"  with the treatment. "1" will be always added. The default is no interaction, i.e., formula = formula(~1).}

\item{Z.hat}{treatment/propensity score prediction from \code{\link{DRPRcomb}}}

\item{Y1.hat}{outcome prediction among treated units from \code{\link{DRPRcomb}}}

\item{Y0.hat}{outcome prediction among untreated units from \code{\link{DRPRcomb}}}

\item{data}{dataframe containing the variables in the model}
}
\value{
estimates and standard errors
}
\description{
Doubly robust estimator
}
\examples{
\dontrun{
DRPRcomb.rslt <- DRPRcomb(Y=twolevel_data$Y, Z=twolevel_data$Z, interZ=(~ W1),
 X=twolevel_data[, c("X1", "X2", "X3", "W1")], ID=twolevel_data$id, data=twolevel_data)

# with final predictions
DR(Y=twolevel_data$Y, Z=twolevel_data$Z, interZ=(~ W1), Z.hat=DRPRcomb.rslt$Z.hat,
 Y1.hat=DRPRcomb.rslt$Y1.hat, Y0.hat=DRPRcomb.rslt$Y0.hat, data=twolevel_data)

# with predictions from glm with fixed effects of clusters
DR(Y=twolevel_data$Y, Z=twolevel_data$Z, interZ=(~ W1), Z.hat=DRPRcomb.rslt$Z.hats$Ztest.hat_glm_1,
 Y1.hat=DRPRcomb.rslt$Y1.hats$Y1test.hat_glm_1, Y0.hat=DRPRcomb.rslt$Y0.hats$Y0test.hat_glm_1,
 data=twolevel_data)

# with predictions from glm wtih random effects of clusters
}
}
