% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combFun.R
\name{overlap}
\alias{overlap}
\title{Overlap plot}
\usage{
overlap(X, Z, bin = 20, ...)
}
\arguments{
\item{X}{numeric vector (e.g., propensity score logit)}

\item{Z}{binary Treatment indicator, 1 - treatment, 0 - control}

\item{bin}{number of bins for histogram}

\item{...}{additional arguments ...}
}
\value{
overlap plot between treated units and control units
}
\description{
overlap plot between the treated group and untreated group
}
\examples{
overlap(X=twolevel_data$lps, Z=twolevel_data$Z)
}
