% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combFun.R
\name{print.summary.CURobustML}
\alias{print.summary.CURobustML}
\title{Print a coefficient table}
\usage{
\method{print}{summary.CURobustML}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{object of class summary}

\item{digits}{integer indicating the number of decimal places}

\item{signif.stars}{whether to add significant starts}

\item{...}{additional arguments ...}
}
\value{

}
\description{
Print a coefficient table
}
